access_control/serizalizers.py
from rest_framework import serializers
from .models import Role, BussinessElement, AccessRule

class RoleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Role
        fields = ["id", "name", "description"]

class AccessRuleSerializer(serializers.ModelSerializer):
    class Meta:
        model = AccessRule
        fields = ["id", "role", "element", "read_permission", "read_all_permission", "create_permission", "update_permission", "update_all_permission","delete_permission","delete_all_permission"]
        read_only_fields = ["id"]
