requirements.txt
asgiref==3.11.0
attrs==25.4.0
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.3
Django==4.2.26
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
pycparser==2.23
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0

